.evaluateThresholds {DataQualityDashboard}	R Documentation

Internal function to evaluate the data quality checks against given thresholds.

Description

Internal function to evaluate the data quality checks against given thresholds.

Usage

.evaluateThresholds(checkResults, tableChecks, fieldChecks, conceptChecks)

Arguments

checkResults
A dataframe containing the results of the data quality checks

tableChecks
A dataframe containing the table checks

fieldChecks
A dataframe containing the field checks

conceptChecks
A dataframe containing the concept checks

[Package DataQualityDashboard version 2.1.2 Index]