.getCheckId {DataQualityDashboard}	R Documentation

Internal function to define the id of each check.

Description

Internal function to define the id of each check.

Usage

.getCheckId(
  checkLevel,
  checkName,
  cdmTableName,
  cdmFieldName = NA,
  conceptId = NA,
  unitConceptId = NA
)

Arguments

checkLevel
The level of the check. Options are table, field, or concept

checkName
The name of the data quality check

cdmTableName
The name of the CDM data table the quality check is applied to

cdmFieldName
The name of the field in the CDM data table the quality check is applied to

conceptId
The concept id the quality check is applied to

unitConceptId
The unit concept id the quality check is applied to

[Package DataQualityDashboard version 2.1.2 Index]