.processCheck {DataQualityDashboard}	R Documentation

Internal function to send the fully qualified sql to the database and return the numerical result.

Description

Internal function to send the fully qualified sql to the database and return the numerical result.

Usage

.processCheck(
  connection,
  connectionDetails,
  check,
  checkDescription,
  sql,
  outputFolder
)

Arguments

connection
A connection for connecting to the CDM database using the DatabaseConnector::connect(connectionDetails) function.

connectionDetails
A connectionDetails object for connecting to the CDM database.

check
The data quality check

checkDescription
The description of the data quality check

sql
The fully qualified sql for the data quality check

outputFolder
The folder to output logs and SQL files to.

[Package DataQualityDashboard version 2.1.2 Index]