.recordResult {DataQualityDashboard}	R Documentation

Internal function to put the results of each quality check into a dataframe.

Description

Internal function to put the results of each quality check into a dataframe.

Usage

.recordResult(
  result = NULL,
  check,
  checkDescription,
  sql,
  executionTime = NA,
  warning = NA,
  error = NA
)

Arguments

result
The result of the data quality check

check
The data quality check

checkDescription
The description of the data quality check

sql
The fully qualified sql for the data quality check

executionTime
The total time it took to execute the data quality check

warning
Any warnings returned from the server

error
Any errors returned from the server

[Package DataQualityDashboard version 2.1.2 Index]