.runCheck {DataQualityDashboard}	R Documentation

Internal function to run and process each data quality check.

Description

Internal function to run and process each data quality check.

Usage

.runCheck(
  checkDescription,
  tableChecks,
  fieldChecks,
  conceptChecks,
  connectionDetails,
  connection,
  cdmDatabaseSchema,
  vocabDatabaseSchema,
  cohortDatabaseSchema,
  cohortDefinitionId,
  outputFolder,
  sqlOnly
)

Arguments

checkDescription
The description of the data quality check

fieldChecks
A dataframe containing the field checks

conceptChecks
A dataframe containing the concept checks

connectionDetails
A connectionDetails object for connecting to the CDM database

connection
A connection for connecting to the CDM database using the DatabaseConnector::connect(connectionDetails) function.

cdmDatabaseSchema
The fully qualified database name of the CDM schema

vocabDatabaseSchema
The fully qualified database name of the vocabulary schema (default is to set it as the cdmDatabaseSchema)

cohortDatabaseSchema
The schema where the cohort table is located.

cohortDefinitionId
The cohort definition id for the cohort you wish to run the DQD on. The package assumes a standard OHDSI cohort table called 'Cohort'

outputFolder
The folder to output logs and SQL files to

sqlOnly
Should the SQLs be executed (FALSE) or just returned (TRUE)?

tablechecks
A dataframe containing the table checks

[Package DataQualityDashboard version 2.1.2 Index]