.summarizeResults {DataQualityDashboard}	R Documentation

Internal function to summarize the results of the DQD run.

Description

Internal function to summarize the results of the DQD run.

Usage

.summarizeResults(checkResults)

Arguments

checkResults
A dataframe containing the results of the checks after running against the database

[Package DataQualityDashboard version 2.1.2 Index]