.writeResultsToCsv {DataQualityDashboard}	R Documentation

Internal function to write the check results to a csv file.

Description

Internal function to write the check results to a csv file.

Usage

.writeResultsToCsv(
  checkResults,
  csvPath,
  columns = c("checkId", "failed", "passed", "isError", "notApplicable", "checkName",
    "checkDescription", "thresholdValue", "notesValue", "checkLevel", "category",
    "subcategory", "context", "checkLevel", "cdmTableName", "cdmFieldName", "conceptId",
    "unitConceptId", "numViolatedRows", "pctViolatedRows", "numDenominatorRows",
    "executionTime", "notApplicableReason", "error", "queryText"),
  delimiter = ","
)

Arguments

checkResults
A dataframe containing the fully summarized data quality check results

csvPath
The path where the csv file should be written

columns
The columns to be included in the csv file. Default is all columns in the checkResults dataframe.

delimiter
The delimiter for the file. Default is comma.

[Package DataQualityDashboard version 2.1.2 Index]