.writeResultsToTable {DataQualityDashboard}	R Documentation

Internal function to write the check results to a table in the database. Requires write access to the database

Description

Internal function to write the check results to a table in the database. Requires write access to the database

Usage

.writeResultsToTable(
  connectionDetails,
  resultsDatabaseSchema,
  checkResults,
  writeTableName,
  cohortDefinitionId
)

Arguments

connectionDetails
A connectionDetails object for connecting to the CDM database

resultsDatabaseSchema
The fully qualified database name of the results schema

checkResults
A dataframe containing the fully summarized data quality check results

writeTableName
The name of the table to be written to the database. Default is "dqdashboard_results".

cohortDefinitionId
(OPTIONAL) The cohort definition id for the cohort you wish to run the DQD on. The package assumes a standard OHDSI cohort table called 'Cohort' with the fields cohort_definition_id and subject_id.

[Package DataQualityDashboard version 2.1.2 Index]