listDqChecks {DataQualityDashboard}	R Documentation

List DQ checks

Description

Details on all checks defined by the DataQualityDashboard Package.

Usage

listDqChecks(
  cdmVersion = "5.3",
  tableCheckThresholdLoc = "default",
  fieldCheckThresholdLoc = "default",
  conceptCheckThresholdLoc = "default"
)

Arguments

cdmVersion
The CDM version to target for the data source. By default, 5.3 is used.

tableCheckThresholdLoc
The location of the threshold file for evaluating the table checks. If not specified the default thresholds will be applied.

fieldCheckThresholdLoc
The location of the threshold file for evaluating the field checks. If not specified the default thresholds will be applied.

conceptCheckThresholdLoc
The location of the threshold file for evaluating the concept checks. If not specified the default thresholds will be applied.

[Package DataQualityDashboard version 2.1.2 Index]