reEvaluateThresholds {DataQualityDashboard}	R Documentation

Re-evaluate Thresholds

Description

Re-evaluate an existing DQD result against an updated thresholds file.

Usage

reEvaluateThresholds(
  jsonFilePath,
  outputFolder,
  outputFile,
  tableCheckThresholdLoc = "default",
  fieldCheckThresholdLoc = "default",
  conceptCheckThresholdLoc = "default",
  cdmVersion = "5.3"
)

Arguments

jsonFilePath
Path to the JSON results file generated using the execute function

outputFolder
The folder to output new JSON result file to

outputFile
File to write results JSON object to

tableCheckThresholdLoc
The location of the threshold file for evaluating the table checks. If not specified the default thresholds will be applied.

fieldCheckThresholdLoc
The location of the threshold file for evaluating the field checks. If not specified the default thresholds will be applied.

conceptCheckThresholdLoc
The location of the threshold file for evaluating the concept checks. If not specified the default thresholds will be applied.

cdmVersion
The CDM version to target for the data source. By default, 5.3 is used.

[Package DataQualityDashboard version 2.1.2 Index]