Write JSON Results to SQL Table

Write JSON Results to SQL Table

writeJsonResultsToTable(
  connectionDetails,
  resultsDatabaseSchema,
  jsonFilePath,
  writeTableName = "dqdashboard_results",
  cohortDefinitionId = c(),
  singleTable = FALSE
)

Arguments

connectionDetails

A connectionDetails object for connecting to the CDM database

resultsDatabaseSchema

The fully qualified database name of the results schema

jsonFilePath

Path to the JSON results file generated using the execute function

writeTableName

Name of table in the database to write results to

cohortDefinitionId

If writing results for a single cohort this is the ID that will be appended to the table name

singleTable

If TRUE, writes all results to a single table. If FALSE (default), writes to 3 separate tables by check level (table, field, concept) (NOTE this default behavior will be deprecated in the future)