Execute DQ checks

Execute DQ checks

executeDqChecks(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  vocabDatabaseSchema = cdmDatabaseSchema,
  cdmSourceName,
  numThreads = 1,
  sqlOnly = FALSE,
  outputFolder = "output",
  verboseMode = FALSE,
  writeToTable = TRUE,
  writeTableName = "dqdashboard_results",
  checkLevels = c("TABLE", "FIELD", "CONCEPT"),
  checkNames = c(),
  cohortDefinitionId = c(),
  cohortDatabaseSchema = resultsDatabaseSchema,
  tablesToExclude = c(),
  cdmVersion = "5.3.1",
  tableCheckThresholdLoc = "default",
  fieldCheckThresholdLoc = "default",
  conceptCheckThresholdLoc = "default"
)

Arguments

connectionDetails
A connectionDetails object for connecting to the CDM database

cdmDatabaseSchema
The fully qualified database name of the CDM schema

resultsDatabaseSchema
The fully qualified database name of the results schema

vocabDatabaseSchema
The fully qualified database name of the vocabulary schema (default is to set it as the cdmDatabaseSchema)

cdmSourceName
The name of the CDM data source

numThreads
The number of concurrent threads to use to execute the queries

sqlOnly
Should the SQLs be executed (FALSE) or just returned (TRUE)?

outputFolder
The folder to output logs and SQL files to

verboseMode
Boolean to determine if the console will show all execution steps. Default = FALSE

writeToTable
Boolean to indicate if the check results will be written to the dqdashboard_results table in the resultsDatabaseSchema. Default is TRUE.

checkLevels
Choose which DQ check levels to execute. Default is all 3 (TABLE, FIELD, CONCEPT)

checkNames
(OPTIONAL) Choose which check names to execute. Names can be found in inst/csv/OMOP_CDM_v[cdmVersion]_Check_Desciptions.csv

cohortDefinitionId
The cohort definition id for the cohort you wish to run the DQD on. The package assumes a standard OHDSI cohort table called 'Cohort' with the fields cohort_definition_id and subject_id.

cohortDatabaseSchema
The schema where the cohort table is located.

tablesToExclude
(OPTIONAL) Choose which CDM tables to exclude from the execution.

cdmVersion
The CDM version to target for the data source. By default, 5.3.1 is used.

tableCheckThresholdLoc
The location of the threshold file for evaluating the table checks. If not specified the default thresholds will be applied.

fieldCheckThresholdLoc
The location of the threshold file for evaluating the field checks. If not specified the default thresholds will be applied.

conceptCheckThresholdLoc
The location of the threshold file for evaluating the concept checks. If not specified the default thresholds will be applied.

Value

If sqlOnly = FALSE, a list object of results