Execute DQ checks

Execute DQ checks

executeDqChecks(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  cdmSourceName,
  numThreads = 1,
  sqlOnly = FALSE,
  outputFolder = "output",
  verboseMode = FALSE,
  writeToTable = TRUE,
  checkLevels = c("TABLE", "FIELD", "CONCEPT"),
  checkNames = c(),
  tablesToExclude = c(),
  cdmVersion = "5.3.1"
)

Arguments

connectionDetails
A connectionDetails object for connecting to the CDM database

cdmDatabaseSchema
The fully qualified database name of the CDM schema

resultsDatabaseSchema
The fully qualified database name of the results schema

cdmSourceName
The name of the CDM data source

numThreads
The number of concurrent threads to use to execute the queries

sqlOnly
Should the SQLs be executed (FALSE) or just returned (TRUE)?

outputFolder
The folder to output logs and SQL files to

verboseMode
Boolean to determine if the console will show all execution steps. Default = FALSE

writeToTable
Boolean to indicate if the check results will be written to the dqdashboard_results table in the resultsDatabaseSchema. Default is TRUE.

checkLevels
Choose which DQ check levels to execute. Default is all 3 (TABLE, FIELD, CONCEPT)

checkNames
(OPTIONAL) Choose which check names to execute. Names can be found in inst/csv/OMOP_CDM_v[cdmVersion]_Check_Desciptions.csv

tablesToExclude
(OPTIONAL) Choose which CDM tables to exclude from the execution.

cdmVersion
The CDM version to target for the data source. By default, 5.3.1 is used.

Value

If sqlOnly = FALSE, a list object of results