writeJsonResultsToTable {DataQualityDashboard}	R Documentation

Write JSON Results to SQL Table

Description

Write JSON Results to SQL Table

Usage

writeJsonResultsToTable(
  connectionDetails,
  resultsDatabaseSchema,
  jsonFilePath,
  writeTableName = "dqdashboard_results",
  cohortDefinitionId = c()
)

Arguments

connectionDetails
A connectionDetails object for connecting to the CDM database

resultsDatabaseSchema
The fully qualified database name of the results schema

jsonFilePath
Path to the JSON results file generated using the execute function

writeTableName
Name of table in the database to write results to

cohortDefinitionId
If writing results for a single cohort this is the ID that will be appended to the table name

[Package DataQualityDashboard version 2.1.2 Index]